change group_size format to uint64_t

+++ src/common/transformations/include/ov_ops/dynamic_quantize.hpp
@@ -22,20 +22,20 @@ public:
     /// \param data Input tensor with data
     /// \param group_sizes Group sizes for dynamic quantization
     /// \param dt_scale Data type for scale output
-    DynamicQuantize(const Output<Node>& data, std::vector<size_t> group_sizes, element::Type dt_scale);
+    DynamicQuantize(const Output<Node>& data, std::vector<uint64_t> group_sizes, element::Type dt_scale);
 
     void validate_and_infer_types() override;
 
     std::shared_ptr<Node> clone_with_new_inputs(const ov::OutputVector& new_args) const override;
-    const std::vector<size_t>& get_group_sizes() const {
+    const std::vector<uint64_t>& get_group_sizes() const {
         return m_group_sizes;
     };
     static std::vector<ov::PartialShape> shape_infer(const DynamicQuantize* op,
                                                      const std::vector<ov::PartialShape>& input_shapes,
-                                                     const std::vector<size_t>& group_sizes);
+                                                     const std::vector<uint64_t>& group_sizes);
 
 private:
-    std::vector<size_t> m_group_sizes;
+    std::vector<uint64_t> m_group_sizes;
     element::Type m_dt_scale;
 };
